Pulling out useMediaQuery into utils
